examples: Wait earlier for command-buffer fence to prevent semaphore reuse
This resolves `VUID-vkAcquireNextImageKHR-semaphore-01779` where
`acquire_next_image()` is asked to signal a semaphore that may still be
waited on by the currently-running submit.

Since the next submit is reusing the existing command buffer and
resources as well it must inevitably wait on a fence from the previous
frame/submit, pulling out that fence to wait on it is the most practical
approach despite teaching bad/slow practices.

Note that this fence was also used to "reuse" the setup command
buffer in the `texture` sample; this was instead generalized to use
another command buffer without waiting for and signaling any fence.